fix(profile) code review fixes
